requirements.txt
django
django-admin
djangorestframework
requests
